% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absence}
\alias{absence}
\title{Estimating the probability of presence in a Psychology experiment}
\format{
A data frame with 13 rows and 6 variables:
\describe{
\item{day}{day of the week}
\item{inscription}{inscription mode}
\item{reminder}{whether a reminder email was sent}
\item{absence}{the number of absent participants}
\item{presence}{the number of present participants}
\item{total}{the total number of participants}
}
}
\source{
Home-made dataset.
}
\usage{
data(absence)
}
\description{
A dataset containing the presence percentage according to weekdays and whether
a reminder email was sent.
}
\keyword{datasets}
