% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absence_multilevel}
\alias{absence_multilevel}
\title{Estimating the probability of presence in a Psychology experiment}
\format{
A data frame with 20 rows and 5 variables:
\describe{
\item{reminder}{whether a reminder email was sent}
\item{researcher}{id of the researcher who collected the data}
\item{presence}{the number of present participants}
\item{absence}{the number of absent participants}
\item{total}{the total number of participants}
}
}
\source{
Home-made dataset.
}
\usage{
data(absence_multilevel)
}
\description{
A dataset containing the presence percentage according to weekdays and the
researcher that collected the data.
}
\keyword{datasets}
