% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admission}
\alias{admission}
\title{Admission data...}
\format{
A data frame with 12 rows and 5 variables:
\describe{
\item{dept}{...}
\item{gender}{...}
\item{admit}{...}
\item{reject}{...}
\item{applications}{...}
}
}
\source{
Adapted from...
}
\usage{
data(admission)
}
\description{
Admission data...
}
\keyword{datasets}
