% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apples}
\alias{apples}
\title{Apples growth}
\format{
A data frame with 480 rows and 5 variables:
\describe{
\item{tree}{id of the tree}
\item{apple}{id of the apple (per apple)}
\item{id}{id for each tree-apple combination}
\item{time}{time in days}
\item{diam}{diameter (in cm)}
}
}
\source{
Home-made dataset.
}
\usage{
data(apples)
}
\description{
A dataset containing the diameter of apples recorded over multiple days and
multiple trees.
}
\keyword{datasets}
