% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_install.R
\name{check_install}
\alias{check_install}
\title{Check the correct installation of the packages needed for the course}
\usage{
check_install(...)
}
\arguments{
\item{...}{The function takes no argument at the moment.}
}
\value{
Returns the status of the installation.
}
\description{
This function can be used without argument to check the installation of the
correct \code{R} version, the install of \code{brms} and \code{rstan} and
to check that \code{Stan} models can be correctly fitted using a simple example.
}
\examples{
\dontrun{
check_install()
}
}
