% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpanzees}
\alias{chimpanzees}
\title{Chimpanzees data...}
\format{
A data frame with 504 rows and 8 variables:
\describe{
\item{actor}{...}
\item{recipient}{...}
\item{condition}{...}
\item{block}{...}
\item{trial}{...}
\item{prosoc_left}{...}
\item{chose_prosoc}{...}
\item{pulled_left}{...}
}
}
\source{
Adapted from the \code{rethinking} package.
}
\usage{
data(chimpanzees)
}
\description{
Chimpanzees data...
}
\keyword{datasets}
