% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fd_nbins}
\alias{fd_nbins}
\title{An implementation of Freedman-Diaconis' rule}
\usage{
fd_nbins(samples)
}
\arguments{
\item{samples}{Numeric, samples from some distribution.}
}
\value{
Returns the number of bins according to Freedman-Diaconis' rule.
}
\description{
An implementation of Freedman-Diaconis' rule to find the "best" number
of bins for a histogram.
}
\examples{
\dontrun{
samples <- rnorm(n = 1e3, mean = 0, sd = 1)
fd_nbins(samples = samples)
}
}
