% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode of a distribution from its samples}
\usage{
find_mode(samples, ...)
}
\arguments{
\item{samples}{Numeric, samples from some distribution.}

\item{...}{Extra parameters to be passed to the \code{stats::density()} function.}
}
\value{
Returns the mode of the distribution.
}
\description{
Find the mode of a distribution from its samples.
}
\examples{
\dontrun{
samples <- rnorm(n = 1e3, mean = 0, sd = 1)
find_mode(samples = samples)
}
}
