% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Meta-analysis data...}
\format{
A data frame with 32 rows and 4 variables:
\describe{
\item{study}{...}
\item{experiment}{...}
\item{yi}{...}
\item{vi}{...}
}
}
\source{
...
}
\usage{
data(meta)
}
\description{
Meta-analysis data...
}
\keyword{datasets}
