% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk}
\alias{milk}
\title{Primate milk data}
\format{
A data frame with 29 rows and 8 variables:
\describe{
\item{clade}{...}
\item{species}{...}
\item{kcal.per.g}{...}
\item{perc.fat}{...}
\item{perc.protein}{...}
\item{perc.lactose}{...}
\item{mass}{...}
\item{neocortex.perc}{...}
}
}
\source{
Adapted from the \code{rethinking} package.
}
\usage{
data(milk)
}
\description{
Small dataset from Hinde and Milligan (2011) on primate milk composition.
This data is discussed in detail in McElreath (2020).
}
\keyword{datasets}
