% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morale}
\alias{morale}
\title{How moral was it?}
\format{
A data frame with 9930 rows and 7 variables:
\describe{
\item{response}{how moral was it? response from 1 to 5}
\item{id}{id of the subject}
\item{age}{age of the subject}
\item{male}{gender indicator}
\item{action}{action condition (binary)}
\item{intention}{intention condition (binary)}
\item{contact}{contact conditio (binary)}
}
}
\source{
Home-made dataset.
}
\usage{
data(morale)
}
\description{
A dataset containing morality judgements in multiple conditions.
}
\keyword{datasets}
