% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_data}
\alias{open_data}
\title{Open data}
\usage{
open_data(df)
}
\arguments{
\item{df}{The desired dataset (from a R package).}
}
\value{
Returns the desired dataset in a dataframe.
}
\description{
Opening data from \code{R} packages.
}
\examples{
\dontrun{
open_data(absence)
}
}
