% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popular}
\alias{popular}
\title{Popularity data...}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
\item{pupil}{...}
\item{school}{...}
\item{popular}{...}
\item{sex}{...}
\item{texp}{...}
\item{teachpop}{...}
}
}
\source{
...
}
\usage{
data(popular)
}
\description{
Popularity data...
}
\keyword{datasets}
