% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plot.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Plotting posterior samples in the style of the \code{BEST} package}
\usage{
posterior_plot(
  samples,
  credmass = 0.91,
  usemode = FALSE,
  compval = NULL,
  rope = NULL,
  showcurve = FALSE,
  maincolour = "steelblue",
  compvalcolour = "darkgreen",
  ROPEcolour = "darkred",
  textsize = 5,
  nbins = NULL
)
}
\arguments{
\item{samples}{Numeric, samples from some distribution.}

\item{credmass}{Numeric, credibility mass (default to 0.91).}

\item{usemode}{Logical, indicating whether we should use the using the mean
(default) or the mode?}

\item{compval}{Numeric, to what value comparing the posterior?}

\item{rope}{Numeric, defining the region of practical equivalence
(such as c(-0.1, 0.1) ).}

\item{showcurve}{Logical, indicates whether we should plot a density (TRUE)
instead of the histogram (FALSE).}

\item{maincolour}{Character string indicating the colour of the
histogram (or density).}

\item{compvalcolour}{Character string indicating the colour the comparison
value's text.}

\item{ROPEcolour}{Character string indicating the colour of the ROPE's text}

\item{textsize}{Numeric, defining the size of the text elements.}

\item{nbins}{Numeric, defining the number of bins.}
}
\value{
A ggplot histogram or density of posterior samples.
}
\description{
Plotting posterior samples in the style of the \code{BEST} package.
}
\examples{
\dontrun{
# getting samples for a normal distribution
samples <- rnorm(n = 1e3, mean = 0, sd = 1)

# plotting it
posterior_plot(
  samples = samples, credmass = 0.96, usemode = FALSE,
  compval = 1, rope = c(0.8, 1.2), showcurve = FALSE
  ) +
  # the resulting plot is a ggplot than can be customised
  labs(x = expression(theta) )
}
}
