% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robot}
\alias{robot}
\title{How long does it take to get a coffee?}
\format{
A data frame with 200 rows and 3 variables:
\describe{
\item{cafe}{id for the place}
\item{afternoon}{morning or afternoon}
\item{wait}{waiting time in minutes}
}
}
\source{
Adapted from the \verb{rethinking package}.
}
\usage{
data(robot)
}
\description{
A dataset containing the time it takes to get a coffee at various coffee
places.
}
\keyword{datasets}
