% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rugged}
\alias{rugged}
\title{Rugged data...}
\format{
A data frame with 234 rows and 5 variables:
\describe{
\item{isocode}{...}
\item{country}{...}
\item{rugged}{...}
\item{cont_africa}{...}
\item{rgdppc_2000}{...}
}
}
\source{
...
}
\usage{
data(tulips)
}
\description{
Rugged data...
}
\keyword{datasets}
