% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tartine1}
\alias{tartine1}
\title{Did the toast fall on the butter side?}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{trial}{the trial number}
\item{side}{the side of the toast, 1 is butter}
}
}
\source{
Home-made dataset.
}
\usage{
data(tartine1)
}
\description{
A dataset containing the results of 100 toast throws.
}
\keyword{datasets}
