% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tartine2}
\alias{tartine2}
\title{Did the toast fall on the butter side?}
\format{
A data frame with 500 rows and 2 variables:
\describe{
\item{trial}{the trial number}
\item{side}{the side of the toast, 1 is butter}
}
}
\source{
home-made dataset
}
\usage{
tartine2
}
\description{
A dataset containing the results of 500 trials.
}
\keyword{datasets}
