% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Survival of Titanic's passengers}
\format{
A data frame with 539 rows and 5 variables:
\describe{
\item{survival}{binary numerical variable (0 or 1)}
\item{pclass}{lower or upper class}
\item{gender}{binary sex indicator variable (female or male)}
\item{age}{age in years}
\item{parch}{number of relatives on board}
}
}
\source{
Adapted from \code{datasets::Titanic}.
}
\usage{
data(titanic)
}
\description{
Information on the survival of the passengers of the Titanic according to
economic status, sex, and age.
}
\keyword{datasets}
