% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tulips}
\alias{tulips}
\title{Tulips data...}
\format{
A data frame with 29 rows and 8 variables:
\describe{
\item{bed}{...}
\item{water}{...}
\item{shade}{...}
\item{blooms}{...}
}
}
\source{
Adapted from the \code{rethinking} package.
}
\usage{
data(tulips)
}
\description{
Tulips data...
}
\keyword{datasets}
