% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apples}
\alias{apples}
\title{Growth of apples}
\format{
A data frame with 480 rows and 5 variables:
\describe{
\item{tree}{id of the ree}
\item{apple}{id of the apple (per apple)}
\item{id}{unique tree-apple id}
\item{time}{time (in days)}
\item{diam}{diameter (in cms)}
}
}
\source{
home-made dataset
}
\usage{
data(apples)
}
\description{
A dataset containing the diameter of apples recorded over multiple days
}
\keyword{datasets}
