% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_install.R
\name{check_install}
\alias{check_install}
\title{Checking the correct installation of the packages needed for the course}
\usage{
check_install(...)
}
\arguments{
\item{...}{The function takes no argument.}
}
\value{
Returns the status of the installation.
}
\description{
This function can be used without argument to check the installation of the
correct \code{R} version, the installation of \code{brms}, and to check that \code{brms}
models can be correctly fitted using a simple example.
}
\examples{
\dontrun{
check_install()
}
}
