% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_install.R
\name{check_install}
\alias{check_install}
\title{Check the correct installation of the packages needed for the course}
\usage{
check_install(...)
}
\arguments{
\item{...}{the function takes no argument at the moment}
}
\value{
returns the status of the installation
}
\description{
Check the correct installation of the packages needed for the course
}
\examples{
\dontrun{
check_install()
}
}
