% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mode}
\alias{find_mode}
\title{Find the mode of a distribution from its samples}
\usage{
find_mode(samples, ...)
}
\arguments{
\item{samples}{samples from some distribution}

\item{...}{extra parameters to be passed to the \code{density} function}
}
\value{
returns the mode of the distribution
}
\description{
Find the mode of a distribution from its samples
}
\examples{
\dontrun{
find_mode(sample = rnorm(n = 1e4, mean = 0, sd = 1) )
}
}
