% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howell}
\alias{howell}
\title{Howell !Kung demography data}
\format{
A data frame with 544 rows and 4 variables:
\describe{
\item{height}{height in cm}
\item{weight}{weight in kg}
\item{age}{age in years}
\item{male}{gender indicator}
}
}
\source{
Adapted from the \code{rethinking} package.
}
\usage{
data(howell)
}
\description{
Demographic data from Kalahari !Kung San people collected by Nancy Howell
}
\keyword{datasets}
