% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk}
\alias{milk}
\title{Waffle House and marriage statistics}
\format{
A data frame with 29 rows and 8 variables:
\describe{
\item{clade}{...}
\item{species}{...}
\item{kcal.per.g}{...}
\item{perc.fat}{...}
\item{perc.protein}{...}
\item{perc.lactose}{...}
\item{mass}{...}
\item{neocortex.perc}{...}
}
}
\source{
...
}
\usage{
data(milk)
}
\description{
...
}
\keyword{datasets}
