% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morale}
\alias{morale}
\title{How moral was it?}
\format{
A data frame with 9930 rows and 7 variables:
\describe{
\item{response}{...}
\item{id}{...}
\item{age}{...}
\item{male}{...}
\item{action}{...}
\item{intention}{...}
\item{contact}{...}
}
}
\source{
home-made dataset
}
\usage{
data(morale)
}
\description{
A dataset containing morality judgments
}
\keyword{datasets}
