% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_slides}
\alias{open_slides}
\title{Open slides in browser}
\usage{
open_slides(cours)
}
\arguments{
\item{cours}{Course number, from 01 to 10.}
}
\value{
Returns nothing but opens the slides in browser.
}
\description{
Opening the slides of the corresponding course (from 01 to 10) in browser.
}
\examples{
\dontrun{
open_slides(cours = 01)
}
}
