% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_slides}
\alias{open_slides}
\title{Open slides in browser}
\usage{
open_slides(cours)
}
\arguments{
\item{cours}{course number}
}
\value{
nothing but opens the slides in browser
}
\description{
Open slides in browser
}
\examples{
\dontrun{
open_slides(cours = 01)
}
}
