% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parents}
\alias{parents}
\title{Inheritance of human height}
\format{
A data frame with 40 rows and 4 variables:
\describe{
\item{gender}{gender indicator variable (M vs. F)}
\item{height}{height in inches}
\item{mother}{height of mother in inches.}
\item{father}{height of father in inches}
}
}
\source{
Adapted from the \code{rethinking} package.
}
\usage{
data(parents)
}
\description{
A dataset containing height data for individuals and their parents.
}
\keyword{datasets}
