% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parents}
\alias{parents}
\title{Height...}
\format{
A data frame with 40 rows and 4 variables:
\describe{
\item{gender}{...}
\item{height}{...}
\item{mother}{...}
\item{father}{...}
}
}
\source{
home-made dataset
}
\usage{
data(parents)
}
\description{
A dataset containing height data...
}
\keyword{datasets}
