% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_plot.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Plotting posterior samples in Kruschke's style}
\usage{
posterior_plot(samples, credmass = 0.89, compval = NULL, rope = NULL)
}
\arguments{
\item{samples}{samples from some distribution}

\item{credmass}{credibility mass (default to 0.89)}

\item{compval}{to what value comparing the posterior?}

\item{rope}{region of practical equivalence (such as c(-0.1, 0.1) )}
}
\value{
a ggplot of some (posterior) samples
}
\description{
Plotting posterior samples in Kruschke's style
}
\examples{
\dontrun{
samples <- rnorm(n = 1e3, mean = 0, sd = 1)
posterior_plot(samples) + labs(x = expression(theta) )
}
}
