% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic data...}
\format{
A data frame with 539 rows and 5 variables:
\describe{
\item{survival}{...}
\item{pclass}{...}
\item{gender}{...}
\item{age}{...}
\item{parch}{...}
}
}
\source{
home-made dataset
}
\usage{
data(titanic)
}
\description{
Titanic data...
}
\keyword{datasets}
