% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tulips}
\alias{tulips}
\title{Waffle House and marriage statistics}
\format{
A data frame with 29 rows and 8 variables:
\describe{
\item{bed}{...}
\item{water}{...}
\item{shade}{...}
\item{blooms}{...}
}
}
\source{
...
}
\usage{
data(tulips)
}
\description{
...
}
\keyword{datasets}
